refactor: clean errors
