+++ src/helpers/errors/index.ts
@@ -1,0 +1,2 @@
+export * from './user-not-found.error';
+export * from './email-is-taken.error';
