--- src/users/errors/index.ts
@@ -1,4 +1,0 @@
-export * from './user-not-found.error';
-export * from './password-not-match.error';
-export * from './email-is-taken.error';
-export * from './cannot-change-to-same-state.error';
