--- src/users/errors/password-not-match.error.ts
@@ -1,10 +1,0 @@
-import { HttpException, HttpStatus } from '@nestjs/common';
-
-export class PasswordDoesntMatchError extends HttpException {
-  constructor() {
-    super(
-      'User account was found but credentials were not correct.',
-      HttpStatus.FORBIDDEN,
-    );
-  }
-}
